requirements.txt
streamlit
pandas
plotly
python-dotenv
psycopg2
csv
os